Refactor quick tabular MDP
