+++ msdm/core/problemclasses/mdp/__init__.py
@@ -1,6 +1,5 @@
 from msdm.core.problemclasses.mdp.mdp import *
 from msdm.core.problemclasses.mdp.deterministic_shortest_path import *
 from msdm.core.problemclasses.mdp.tabularmdp import *
-from msdm.core.problemclasses.mdp.quicktabularmdp import QuickTabularMDP
-from msdm.core.problemclasses.mdp.quickmdp import QuickMDP
+from msdm.core.problemclasses.mdp.quickmdp import QuickMDP, QuickTabularMDP
 from msdm.core.problemclasses.mdp.policy import *
