--- msdm/core/problemclasses/mdp/quicktabularmdp.py
@@ -1,5 +1,0 @@
-from msdm.core.problemclasses.mdp.tabularmdp import TabularMarkovDecisionProcess
-from msdm.core.problemclasses.mdp.quickmdp import QuickMDP
-
-class QuickTabularMDP(QuickMDP,TabularMarkovDecisionProcess):
-    pass
