+++ msdm/tests/test_core_mdp.py
@@ -96,8 +96,7 @@ def test_TabularMarkovDecisionProcess_from_matrices():
         assert pi1.converged == pi2.converged
 
 def test_QuickMDP_equivalence():
-    from msdm.core.problemclasses.mdp.quickmdp import QuickMDP
-    from msdm.core.problemclasses.mdp.quicktabularmdp import QuickTabularMDP 
+    from msdm.core.problemclasses.mdp.quickmdp import QuickMDP, QuickTabularMDP
 
     for mdp in test_mdps:
         mdp : TabularMarkovDecisionProcess
